Decouple KYC pages
